Fix linter warning
