+++ evals/.tool-versions
@@ -1,4 +1,4 @@
 python 3.13.2
 golang 1.24.2
-rust 1.86.0
+rust 1.85.1
 nodejs 20.18.1
